:art: adjust outlier rejection submodule
